Split out MessageSearch to its own file
